test: the transaction API-related suite re-structuring
