+++ packages/integration-tests/src/playbook/scenarios/transfers/transferETH.ts
@@ -1,8 +1,9 @@
 import * as ethers from "ethers";
+import { promises as fs } from "fs";
 import * as zksync from "zksync-web3";
 
 import { localConfig } from "../../../config";
-import { Logger, Wallets } from "../../../entities";
+import { Buffer, Logger, Wallets } from "../../../entities";
 import { Helper } from "../../../helper";
 
 export const transferEth = async function (sum = "0.000009", address: string = Wallets.mainWalletPrivateKey) {
@@ -11,6 +12,8 @@ export const transferEth = async function (sum = "0.000009", address: string = W
   const ethProvider = ethers.getDefaultProvider(localConfig.L1Network);
   const syncWallet = new zksync.Wallet(localConfig.privateKey, syncProvider, ethProvider);
   const syncWallet2 = new zksync.Wallet(address, syncProvider, ethProvider);
+  const playbookRoot = "src/playbook/";
+  const bufferFile = playbookRoot + Buffer.txEthTransfer;
 
   const transfer = await syncWallet.transfer({
     to: syncWallet2.address,
@@ -20,6 +23,7 @@ export const transferEth = async function (sum = "0.000009", address: string = W
 
   const txHash = transfer.hash;
   await helper.txHashLogger(Logger.transfer, txHash, "ETH");
+  await fs.writeFile(bufferFile, txHash);
 
   return txHash;
 };
