+++ packages/integration-tests/tests/api/transactions/multiCall.test.ts
@@ -28,246 +28,4 @@ describe("Multicall transactions", () => {
     txMulticall = await playbook.useMultiCallContracts();
     expect(txMulticall).toContain(Logger.txHashStartsWith);
   });
-
-  describe("Verify the multicall transactions via /transactions/${txHash}", () => {
-    //@id1464:I --> @id1468
-    it("Verify transaction for the Root contract via /transactions/${txHash}", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txMultiCallRoot);
-      const apiRoute = `/transactions/${txHash}`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ hash: txHash })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ value: "0" })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ isL1Originated: false })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ transactionIndex: 0 })));
-    });
-
-    //@id1465:I --> @id1469
-    it("Verify transaction for the Middle contract via /transactions/${txHash}", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txMultiCallMiddle);
-      const apiRoute = `/transactions/${txHash}`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ hash: txHash })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ value: "0" })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ isL1Originated: false })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ transactionIndex: 0 })));
-    });
-
-    //@id1466:I --> @id1470
-    it("Verify transaction for the Caller contract via /transactions/${txHash}", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txMultiCallCaller);
-      const apiRoute = `/transactions/${txHash}`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ hash: txHash })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ value: "0" })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ isL1Originated: false })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ transactionIndex: 0 })));
-    });
-
-    //@id1471
-    it("Verify transaction for the use multicall contract via /transactions/${txHash}", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txUseMultiCallContracts);
-      const apiRoute = `/transactions/${txHash}`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ hash: txHash })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ value: "0" })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ isL1Originated: false })))
-        .expect((res) => expect(res.body).toStrictEqual(expect.objectContaining({ transactionIndex: 0 })));
-    });
-  });
-
-  describe("Verify the multicall transactions via /transactions/${txHash}/transfers", () => {
-    //@id1472
-    it("Verify transaction for the Root contract via /transactions/${txHash}/transfers", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txMultiCallRoot);
-      const apiRoute = `/transactions/${txHash}/transfers?page=1&limit=10`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ to: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ type: TransactionsType.fee }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ from: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ to: Wallets.richWalletAddress }))
-        )
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ type: TransactionsType.refund }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })));
-    });
-
-    //@id1473
-    it("Verify transaction for the Middle contract via /transactions/${txHash}/transfers", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txMultiCallMiddle);
-      const apiRoute = `/transactions/${txHash}/transfers?page=1&limit=10`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ to: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ type: TransactionsType.fee }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ from: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ to: Wallets.richWalletAddress }))
-        )
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ type: TransactionsType.refund }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })));
-    });
-
-    //@id1474
-    it("Verify transaction for the Caller contract via /transactions/${txHash}/transfers", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txMultiCallCaller);
-      const apiRoute = `/transactions/${txHash}/transfers?page=1&limit=10`;
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ to: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ type: TransactionsType.fee }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ from: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ to: Wallets.richWalletAddress }))
-        )
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ type: TransactionsType.refund }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })));
-    });
-
-    //@id1475
-    it("Verify transaction for the use multicall contract via /transactions/${txHash}/transfers", async () => {
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      txHash = await helper.getStringFromFile(bufferRoute + Buffer.txUseMultiCallContracts);
-
-      const apiRoute = `/transactions/${txHash}/transfers?page=1&limit=10`;
-      await setTimeout(localConfig.standardPause); //works unstable without timeout
-
-      return request(environment.blockExplorerAPI)
-        .get(apiRoute)
-        .expect(200)
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ from: Wallets.richWalletAddress }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ to: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ type: TransactionsType.fee }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ from: Token.ETHER_PULL_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ to: Wallets.richWalletAddress }))
-        )
-        .expect((res) => expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ transactionHash: txHash })))
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ tokenAddress: Token.ETHER_ERC20_Address }))
-        )
-        .expect((res) =>
-          expect(res.body.items[1]).toStrictEqual(expect.objectContaining({ type: TransactionsType.refund }))
-        )
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ fields: null })))
-        .expect((res) => expect(res.body.items[0]).toStrictEqual(expect.objectContaining({ token: null })));
-    });
-  });
 });
